@@ -4,9 +4,9 @@ plugins {
 }
 
 dependencies {
-    testAnnotationProcessor(platform(mn.micronaut.bom))
+    testAnnotationProcessor(platform(mn.micronaut.core.bom))
     testAnnotationProcessor(mn.micronaut.inject.java)
-    testImplementation(platform(mn.micronaut.bom))
+    testImplementation(platform(mn.micronaut.core.bom))
     testImplementation(mn.micronaut.inject.java)
 
     testImplementation(mn.micronaut.http.client)
@@ -17,8 +17,8 @@ dependencies {
 }
 
 java {
-    sourceCompatibility = JavaVersion.toVersion("1.8")
-    targetCompatibility = JavaVersion.toVersion("1.8")
+    sourceCompatibility = JavaVersion.toVersion("17")
+    targetCompatibility = JavaVersion.toVersion("17")
 }
 
 tasks.named("test") {
