@@ -85,6 +85,7 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
     private final HttpParameters parameters;
     private final String path;
     private final ContainerConfig config;
+    private final ConversionService conversionService;
     private Cookies cookies;
     private MicronautAwsProxyResponse<?> response;
     private T decodedBody;
@@ -92,29 +93,31 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
     /**
      * Default constructor.
      *
-     * @param path            The path
-     * @param awsProxyRequest The underlying request
-     * @param securityContext The {@link SecurityContext}
-     * @param lambdaContext   The lambda context
-     * @param config          The container configuration
+     * @param path              The path
+     * @param awsProxyRequest   The underlying request
+     * @param securityContext   The {@link SecurityContext}
+     * @param lambdaContext     The lambda context
+     * @param config            The container configuration
+     * @param conversionService The conversion service
      */
     MicronautAwsProxyRequest(
             String path,
             AwsProxyRequest awsProxyRequest,
             SecurityContext securityContext,
             Context lambdaContext,
-            ContainerConfig config) {
+            ContainerConfig config,
+            ConversionService conversionService) {
         this.config = config;
         this.awsProxyRequest = awsProxyRequest;
         this.path = path;
+        this.conversionService = conversionService;
         final String httpMethod = awsProxyRequest.getHttpMethod();
         this.httpMethod = StringUtils.isNotEmpty(httpMethod) ? HttpMethod.valueOf(httpMethod) : HttpMethod.GET;
-
         this.headers = awsProxyRequest.getMultiValueHeaders() != null || awsProxyRequest.getHeaders() != null ?
-            new AwsHeaders(awsProxyRequest.getMultiValueHeaders(), awsProxyRequest.getHeaders()) : new SimpleHttpHeaders(ConversionService.SHARED);
+            new AwsHeaders(awsProxyRequest.getMultiValueHeaders(), awsProxyRequest.getHeaders(), conversionService) : new SimpleHttpHeaders(conversionService);
 
         final MultiValuedTreeMap<String, String> params = awsProxyRequest.getMultiValueQueryStringParameters();
-        this.parameters = params != null ? new AwsParameters() : new SimpleHttpParameters(ConversionService.SHARED);
+        this.parameters = params != null ? new AwsParameters() : new SimpleHttpParameters(conversionService);
 
         final AwsProxyRequestContext requestContext = awsProxyRequest.getRequestContext();
         setAttribute(API_GATEWAY_CONTEXT_PROPERTY, requestContext);
@@ -139,8 +142,7 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
         if (securityContext == null) {
             return false;
         }
-        if (securityContext instanceof AwsProxySecurityContext) {
-            AwsProxySecurityContext awsProxySecurityContext = (AwsProxySecurityContext) securityContext;
+        if (securityContext instanceof AwsProxySecurityContext awsProxySecurityContext) {
             if (awsProxySecurityContext.getEvent() == null ||
                     awsProxySecurityContext.getEvent().getRequestContext() == null ||
                     awsProxySecurityContext.getEvent().getRequestContext().getIdentity() == null) {
@@ -184,7 +186,7 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
     @NonNull
     public Cookies getCookies() {
         if (cookies == null) {
-            SimpleCookies simpleCookies = new SimpleCookies(ConversionService.SHARED);
+            SimpleCookies simpleCookies = new SimpleCookies(conversionService);
             getHeaders().getAll(HttpHeaders.COOKIE).forEach(cookieValue -> {
                 List<HeaderValue> parsedHeaders = parseHeaderValue(cookieValue, ";", ",");
 
@@ -312,7 +314,7 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
     @NonNull
     public <T1> Optional<T1> getBody(Argument<T1> type) {
         if (decodedBody != null) {
-            return ConversionService.SHARED.convert(decodedBody, type);
+            return conversionService.convert(decodedBody, type);
         }
         final String body = awsProxyRequest.getBody();
         if (body != null) {
@@ -321,13 +323,13 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
                 if (type.getType().isInstance(bytes)) {
                     return (Optional<T1>) Optional.of(bytes);
                 }
-                return ConversionService.SHARED.convert(bytes, type);
+                return conversionService.convert(bytes, type);
             }
             if (type.getType().isInstance(body)) {
                 return (Optional<T1>) Optional.of(body);
             } else {
                 final byte[] bytes = body.getBytes(getCharacterEncoding());
-                return ConversionService.SHARED.convert(bytes, type);
+                return conversionService.convert(bytes, type);
             }
         }
         return Optional.empty();
@@ -495,7 +497,7 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
         public <T> Optional<T> get(CharSequence name, ArgumentConversionContext<T> conversionContext) {
             final String v = get(name);
             if (v != null) {
-                return ConversionService.SHARED.convert(v, conversionContext);
+                return conversionService.convert(v, conversionContext);
             }
             return Optional.empty();
         }
@@ -507,8 +509,10 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
     private static class AwsHeaders implements HttpHeaders {
 
         private final Map<String, List<String>> headers;
+        private final ConversionService conversionService;
 
-        public AwsHeaders(@Nullable Headers multivalueHeaders, @Nullable SingleValueHeaders singleValueHeaders) {
+        public AwsHeaders(@Nullable Headers multivalueHeaders, @Nullable SingleValueHeaders singleValueHeaders, ConversionService conversionService) {
+            this.conversionService = conversionService;
             if (multivalueHeaders == null && singleValueHeaders == null) {
                 headers = Collections.emptyMap();
             } else {
@@ -567,7 +571,7 @@ public class MicronautAwsProxyRequest<T> implements HttpRequest<T> {
         public <T> Optional<T> get(CharSequence name, ArgumentConversionContext<T> conversionContext) {
             final String v = get(name);
             if (v != null) {
-                return ConversionService.SHARED.convert(v, conversionContext);
+                return conversionService.convert(v, conversionContext);
             }
             return Optional.empty();
         }
