@@ -32,70 +32,44 @@ import io.micronaut.core.annotation.AnnotationMetadata;
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.annotation.Nullable;
 import io.micronaut.core.annotation.TypeHint;
-import io.micronaut.core.async.publisher.Publishers;
-import io.micronaut.core.async.subscriber.CompletionAwareSubscriber;
 import io.micronaut.core.bind.BeanPropertyBinder;
 import io.micronaut.core.convert.exceptions.ConversionErrorException;
 import io.micronaut.core.convert.value.ConvertibleValues;
 import io.micronaut.core.convert.value.ConvertibleValuesMap;
 import io.micronaut.core.type.Argument;
-import io.micronaut.core.type.TypeVariableResolver;
 import io.micronaut.core.util.ArgumentUtils;
 import io.micronaut.core.util.CollectionUtils;
-import io.micronaut.core.util.StringUtils;
 import io.micronaut.function.aws.HandlerUtils;
 import io.micronaut.function.aws.LambdaApplicationContextBuilder;
 import io.micronaut.http.HttpAttributes;
-import io.micronaut.http.HttpMethod;
-import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.HttpStatus;
 import io.micronaut.http.MediaType;
-import io.micronaut.http.MutableHttpResponse;
 import io.micronaut.http.annotation.Body;
-import io.micronaut.http.annotation.Consumes;
 import io.micronaut.http.annotation.Produces;
 import io.micronaut.http.annotation.Status;
 import io.micronaut.http.bind.RequestBinderRegistry;
-import io.micronaut.http.context.ServerRequestContext;
 import io.micronaut.http.server.HttpServerConfiguration;
 import io.micronaut.http.server.RouteExecutor;
 import io.micronaut.http.server.binding.RequestArgumentSatisfier;
-import io.micronaut.http.server.exceptions.response.ErrorContext;
 import io.micronaut.http.server.exceptions.response.ErrorResponseProcessor;
 import io.micronaut.inject.qualifiers.Qualifiers;
 import io.micronaut.jackson.codec.JsonMediaTypeCodec;
 import io.micronaut.scheduling.executor.ExecutorSelector;
-import io.micronaut.web.router.MethodBasedRouteMatch;
 import io.micronaut.web.router.RouteInfo;
-import io.micronaut.web.router.RouteMatch;
 import io.micronaut.web.router.Router;
-import io.micronaut.web.router.UriRouteMatch;
 import io.micronaut.web.router.resource.StaticResourceResolver;
-import org.reactivestreams.Publisher;
-import org.reactivestreams.Subscriber;
-import org.reactivestreams.Subscription;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
 
 import java.io.Closeable;
 import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.TreeMap;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiFunction;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 /**
  * Main entry for AWS API proxy with Micronaut.
@@ -115,9 +89,7 @@ import java.util.stream.Collectors;
                 CognitoAuthorizerClaims.class,
                 ContainerConfig.class,
                 ErrorModel.class,
-                SingleValueHeaders.class,
                 Headers.class,
-                TreeMap.class,
                 MultiValuedTreeMap.class,
                 AwsProxySecurityContext.class
         }
@@ -137,7 +109,7 @@ public final class MicronautLambdaContainerHandler
     private Router router;
     private ErrorResponseProcessor errorResponseProcessor;
     private RouteExecutor routeExecutor;
-    private final Map<MediaType, BiFunction<Argument<?>, String, Optional<Object>>> mediaTypeBodyDecoder = new HashMap<>();
+    final Map<MediaType, BiFunction<Argument<?>, String, Optional<Object>>> mediaTypeBodyDecoder = new HashMap<>();
 
     /**
      * Default constructor.
@@ -325,79 +297,19 @@ public final class MicronautLambdaContainerHandler
         HandlerUtils.configureWithContext(this, lambdaContext);
 
         try {
-            ServerRequestContext.with(containerRequest, () -> {
-                Optional<UriRouteMatch> routeMatch = containerRequest.getAttribute(HttpAttributes.ROUTE_MATCH, UriRouteMatch.class);
-
-                if (!routeMatch.isPresent()) {
-                    handlePossibleErrorStatus(containerRequest, containerResponse);
-                    return;
-                }
-
-                handleRouteMatch(routeMatch.get(), containerRequest, containerResponse);
-            });
+            new AwsRequestLifecycle(this, routeExecutor, containerRequest, containerResponse).run()
+                .onComplete((r, t) -> {
+                    if (t == null) {
+                        toAwsProxyResponse(containerResponse, r);
+                    } else {
+                        // should never happen, it should be transformed to an error response with onError
+                        LOG.warn("Error in response flow");
+                    }
+                    containerResponse.close();
+                });
         } finally {
             Timer.stop(TIMER_REQUEST);
         }
-    }
-
-    private void handleRouteMatch(
-            RouteMatch<?> originalRoute,
-            MicronautAwsProxyRequest<?> request,
-            MicronautAwsProxyResponse<?> response
-    ) {
-        final AnnotationMetadata annotationMetadata = originalRoute.getAnnotationMetadata();
-        annotationMetadata.stringValue(Produces.class)
-                .map(MediaType::new)
-                .ifPresent(response::contentType);
-
-        Flux<MutableHttpResponse<?>> routeResponse;
-        try {
-            decodeRequestBody(request, originalRoute)
-                .ifPresent(((MicronautAwsProxyRequest) request)::setDecodedBody
-            );
-
-            RouteMatch<?> route = requestArgumentSatisfier.fulfillArgumentRequirements(originalRoute, request, false);
-
-            Flux<RouteMatch<?>> routeMatchPublisher = Flux.just(route);
-
-            routeResponse = routeExecutor.executeRoute(
-                    request,
-                    true,
-                    routeMatchPublisher
-            ).mapNotNull(r -> convertResponseBody(response, r.getAttribute(HttpAttributes.ROUTE_INFO, RouteInfo.class).get(), r.body()).block());
-        } catch (Exception e) {
-            routeResponse = Flux.from(routeExecutor.filterPublisher(new AtomicReference<>(request), routeExecutor.onError(e, request)));
-        }
-
-        routeResponse
-                .contextWrite(ctx -> ctx.put(ServerRequestContext.KEY, request))
-                .subscribe(new CompletionAwareSubscriber<HttpResponse<?>>() {
-                    @Override
-                    protected void doOnSubscribe(Subscription subscription) {
-                        subscription.request(1);
-                    }
-
-                    @Override
-                    protected void doOnNext(HttpResponse<?> message) {
-                        toAwsProxyResponse(response, message);
-                        subscription.request(1);
-                    }
-
-                    @Override
-                    protected void doOnError(Throwable throwable) {
-                        try {
-                            final MutableHttpResponse<?> defaultErrorResponse = routeExecutor.createDefaultErrorResponse(request, throwable);
-                            toAwsProxyResponse(response, defaultErrorResponse);
-                        } finally {
-                            response.close();
-                        }
-                    }
-
-                    @Override
-                    protected void doOnComplete() {
-                        response.close();
-                    }
-                });
     }
 
     private MicronautAwsProxyResponse<?> toAwsProxyResponse(
@@ -420,78 +332,6 @@ public final class MicronautLambdaContainerHandler
     private void populateMediaTypeBodyDecoders() {
         mediaTypeBodyDecoder.put(MediaType.APPLICATION_JSON_TYPE, this::getJsonDecodedBody);
         mediaTypeBodyDecoder.put(MediaType.APPLICATION_FORM_URLENCODED_TYPE, this::getFormUrlEncodedDecodedBody);
-    }
-
-    @NonNull
-    private static Optional<String> parseUndecodedBody(@NonNull MicronautAwsProxyRequest<?> containerRequest,
-                                                @NonNull RouteMatch<?> finalRoute,
-                                                @NonNull MediaType mediaType) {
-        if (containerRequest.isBodyDecoded()) {
-            return Optional.empty();
-        }
-        if (!HttpMethod.permitsRequestBody(containerRequest.getMethod())) {
-            return Optional.empty();
-        }
-        final MediaType requestContentType = containerRequest.getContentType().orElse(null);
-        if (requestContentType == null) {
-            return Optional.empty();
-        }
-        if (!mediaType.getExtension().equals(requestContentType.getExtension())) {
-            return Optional.empty();
-        }
-        final MediaType[] expectedContentType = finalRoute.getAnnotationMetadata().getValue(Consumes.class, MediaType[].class).orElse(null);
-        return (expectedContentType == null || Arrays.stream(expectedContentType).anyMatch(ct -> mediaType.getExtension().equals(ct.getExtension()))) ?
-            containerRequest.getBody(String.class) :
-            Optional.empty();
-    }
-
-    @NonNull
-    private Optional<Object> decodeRequestBody(@NonNull MicronautAwsProxyRequest<?> containerRequest,
-                                               @NonNull RouteMatch<?> finalRoute) {
-        return mediaTypeBodyDecoder.entrySet()
-            .stream()
-            .map(e -> decodeRequestBody(containerRequest, finalRoute, e))
-            .filter(Optional::isPresent)
-            .map(Optional::get)
-            .findFirst();
-    }
-
-    private static Optional<Object> decodeRequestBody(@NonNull MicronautAwsProxyRequest<?> containerRequest,
-                                                      @NonNull RouteMatch<?> finalRoute,
-                                                      @NonNull Map.Entry<MediaType, BiFunction<Argument<?>, String, Optional<Object>>> entry) {
-        return parseUndecodedBody(containerRequest, finalRoute, entry.getKey())
-            .flatMap(body -> decodeRequestBody(body, entry.getValue(), finalRoute));
-    }
-
-    @NonNull
-    private static Optional<Object> decodeRequestBody(@NonNull String body,
-                                               @NonNull BiFunction<Argument<?>, String, Optional<Object>> function,
-                                               @NonNull RouteMatch<?> finalRoute) {
-        if (StringUtils.isNotEmpty(body)) {
-            Argument<?> bodyArgument = parseBodyArgument(finalRoute);
-            return function.apply(bodyArgument, body);
-        }
-        return Optional.empty();
-    }
-
-    @Nullable
-    private static Argument<?> parseBodyArgument(@NonNull RouteMatch<?> finalRoute) {
-        Argument<?> bodyArgument = finalRoute.getBodyArgument().orElse(null);
-        if (bodyArgument == null) {
-            if (finalRoute instanceof MethodBasedRouteMatch) {
-                bodyArgument = Arrays.stream(((MethodBasedRouteMatch) finalRoute).getArguments())
-                    .filter(arg -> HttpRequest.class.isAssignableFrom(arg.getType()))
-                    .findFirst()
-                    .flatMap(TypeVariableResolver::getFirstTypeVariable).orElse(null);
-            }
-        }
-        if (bodyArgument != null) {
-            final Class<?> rawType = bodyArgument.getType();
-            if (Publishers.isConvertibleToPublisher(rawType) || HttpRequest.class.isAssignableFrom(rawType)) {
-                bodyArgument = bodyArgument.getFirstTypeVariable().orElse(Argument.OBJECT_ARGUMENT);
-            }
-        }
-        return bodyArgument;
     }
 
     @NonNull
@@ -568,179 +408,11 @@ public final class MicronautLambdaContainerHandler
         }
     }
 
-    private Mono<MutableHttpResponse<?>> convertResponseBody(
-            MicronautAwsProxyResponse<?> containerResponse,
-            RouteInfo<?> routeInfo,
-            Object body) {
-        if (Publishers.isConvertibleToPublisher(body)) {
-            Mono<?> single;
-            if (Publishers.isSingle(body.getClass()) || routeInfo.getReturnType().isSpecifiedSingle()) {
-                single = Mono.from(Publishers.convertPublisher(body, Publisher.class));
-            } else {
-                single = Flux.from(Publishers.convertPublisher(body, Publisher.class)).collectList();
-            }
-            return single.map((Function<Object, MutableHttpResponse<?>>) o -> {
-                if (!(o instanceof MicronautAwsProxyResponse)) {
-                    ((MutableHttpResponse) containerResponse).body(o);
-                }
-                applyRouteConfig(containerResponse, routeInfo);
-                return containerResponse;
-            });
-        } else {
-            if (!(body instanceof MicronautAwsProxyResponse)) {
-                applyRouteConfig(containerResponse, routeInfo);
-                ((MutableHttpResponse) containerResponse).body(body);
-            }
-            return Mono.just(containerResponse);
-        }
-    }
-
     private void applyRouteConfig(MicronautAwsProxyResponse<?> containerResponse, RouteInfo<?> finalRoute) {
         if (!containerResponse.getContentType().isPresent()) {
             finalRoute.getAnnotationMetadata().getValue(Produces.class, String.class).ifPresent(containerResponse::contentType);
         }
         finalRoute.getAnnotationMetadata().getValue(Status.class, HttpStatus.class).ifPresent(httpStatus -> containerResponse.status(httpStatus));
-    }
-
-    private void handlePossibleErrorStatus(
-            MicronautAwsProxyRequest<?> request,
-            MicronautAwsProxyResponse<?> response) {
-
-        final MediaType contentType = request.getContentType().orElse(null);
-        final String requestMethodName = request.getMethodName();
-
-        // if there is no route present try to locate a route that matches a different HTTP method
-        final List<UriRouteMatch<?, ?>> anyMatchingRoutes = router
-                .findAny(request.getPath(), request)
-                .collect(Collectors.toList());
-        final Collection<MediaType> acceptedTypes = request.accept();
-        final boolean hasAcceptHeader = CollectionUtils.isNotEmpty(acceptedTypes);
-
-        Set<MediaType> acceptableContentTypes = contentType != null ? new HashSet<>(5) : null;
-        Set<String> allowedMethods = new HashSet<>(5);
-        Set<MediaType> produceableContentTypes = hasAcceptHeader ? new HashSet<>(5) : null;
-        for (UriRouteMatch<?, ?> anyRoute : anyMatchingRoutes) {
-            final String routeMethod = anyRoute.getRoute().getHttpMethodName();
-            if (!requestMethodName.equals(routeMethod)) {
-                allowedMethods.add(routeMethod);
-            }
-            if (contentType != null && !anyRoute.doesConsume(contentType)) {
-                acceptableContentTypes.addAll(anyRoute.getRoute().getConsumes());
-            }
-            if (hasAcceptHeader && !anyRoute.doesProduce(acceptedTypes)) {
-                produceableContentTypes.addAll(anyRoute.getRoute().getProduces());
-            }
-        }
-
-        if (CollectionUtils.isNotEmpty(acceptableContentTypes)) {
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("Content type not allowed for URI {}, method {}, and content type {}", request.getUri(),
-                        requestMethodName, contentType);
-            }
-
-            handleStatusError(
-                    request,
-                    response,
-                    HttpResponse.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE),
-                    "Content Type [" + contentType + "] not allowed. Allowed types: " + acceptableContentTypes);
-            return;
-        }
-
-        if (CollectionUtils.isNotEmpty(produceableContentTypes)) {
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("Content type not allowed for URI {}, method {}, and content type {}", request.getUri(),
-                        requestMethodName, contentType);
-            }
-
-            handleStatusError(
-                    request,
-                    response,
-                    HttpResponse.status(HttpStatus.NOT_ACCEPTABLE),
-                    "Specified Accept Types " + acceptedTypes + " not supported. Supported types: " + produceableContentTypes);
-            return;
-        }
-
-        if (!allowedMethods.isEmpty()) {
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("Method not allowed for URI {} and method {}", request.getUri(), requestMethodName);
-            }
-
-            handleStatusError(
-                    request,
-                    response,
-                    HttpResponse.notAllowedGeneric(allowedMethods),
-                    "Method [" + requestMethodName + "] not allowed for URI [" + request.getUri() + "]. Allowed methods: " + allowedMethods);
-            return;
-        }
-
-        handleStatusError(
-                request,
-                response,
-                HttpResponse.notFound(),
-                "Page Not Found");
-    }
-
-    private void handleStatusError(
-            MicronautAwsProxyRequest<?> request,
-            MicronautAwsProxyResponse<?> response,
-            MutableHttpResponse<?> defaultResponse,
-            String message) {
-
-        Optional<RouteMatch<Object>> statusRoute = router.findStatusRoute(defaultResponse.status(), request);
-        if (statusRoute.isPresent()) {
-            handleRouteMatch(statusRoute.get(), request, response);
-        } else {
-            if (request.getMethod() != HttpMethod.HEAD) {
-                defaultResponse = errorResponseProcessor.processResponse(
-                        ErrorContext.builder(request).errorMessage(message).build(),
-                        defaultResponse
-                );
-                if (!defaultResponse.getContentType().isPresent()) {
-                    defaultResponse = defaultResponse.contentType(MediaType.APPLICATION_JSON_TYPE);
-                }
-            }
-
-            filterAndEncodeResponse(request, response, Publishers.just(defaultResponse));
-        }
-    }
-
-    private void filterAndEncodeResponse(
-            MicronautAwsProxyRequest<?> request,
-            MicronautAwsProxyResponse<?> response,
-            Publisher<MutableHttpResponse<?>> responsePublisher) {
-        AtomicReference<HttpRequest<?>> requestReference = new AtomicReference<>(request);
-
-        Flux.from(routeExecutor.filterPublisher(requestReference, responsePublisher))
-                .contextWrite(ctx -> ctx.put(ServerRequestContext.KEY, request))
-                .subscribe(new Subscriber<MutableHttpResponse<?>>() {
-                    Subscription subscription;
-                    @Override
-                    public void onSubscribe(Subscription s) {
-                        this.subscription = s;
-                        s.request(1);
-                    }
-
-                    @Override
-                    public void onNext(MutableHttpResponse<?> message) {
-                        toAwsProxyResponse(response, message);
-                        subscription.request(1);
-                    }
-
-                    @Override
-                    public void onError(Throwable t) {
-                        try {
-                            final MutableHttpResponse<?> defaultErrorResponse = routeExecutor.createDefaultErrorResponse(request, t);
-                            toAwsProxyResponse(response, defaultErrorResponse);
-                        } finally {
-                            response.close();
-                        }
-                    }
-
-                    @Override
-                    public void onComplete() {
-                        response.close();
-                    }
-                });
     }
 
     @Override
