@@ -15,11 +15,12 @@
  */
 package io.micronaut.function.aws;
 
-import com.amazonaws.services.lambda.runtime.*;
-import io.micronaut.context.event.ApplicationEventPublisher;
-import io.micronaut.core.annotation.NonNull;
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestHandler;
 import io.micronaut.context.ApplicationContext;
 import io.micronaut.context.ApplicationContextBuilder;
+import io.micronaut.context.event.ApplicationEventPublisher;
+import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.convert.ArgumentConversionContext;
 import io.micronaut.core.convert.ConversionContext;
 import io.micronaut.core.convert.ConversionError;
@@ -29,7 +30,7 @@ import io.micronaut.function.aws.event.AfterExecutionEvent;
 import io.micronaut.function.executor.AbstractFunctionExecutor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.slf4j.MDC;
+
 import java.util.Optional;
 
 /**
@@ -47,9 +48,6 @@ public abstract class MicronautRequestHandler<I, O> extends AbstractFunctionExec
     // See: https://github.com/aws/aws-xray-sdk-java/issues/251
     public static final String LAMBDA_TRACE_HEADER_PROP = "com.amazonaws.xray.traceHeader";
 
-    /**
-     * Logger for the application context creation errors.
-     */
     private static final Logger LOG = LoggerFactory.getLogger(MicronautRequestHandler.class);
 
     @SuppressWarnings("unchecked")
